requirements.txt
asgiref==3.5.0
Django==4.0.3
django-environ==0.8.1
environ==1.0
Pillow==9.1.0
six==1.16.0
sqlparse==0.4.2
